+++ controller/pkg/aclprovider/ipsetprovider_windows.go
@@ -3,8 +3,6 @@
 package provider
 
 import (
-	"fmt"
-
 	"github.com/aporeto-inc/go-ipset/ipset"
 )
 
@@ -27,29 +25,55 @@ type Ipset interface {
 	Test(entry string) (bool, error)
 }
 
-type goIpsetProvider struct{}
+type ipsetProvider struct{}
+
+type winIpSet struct{}
 
 // NewIpset returns an IpsetProvider interface based on the go-ipset
 // external package.
-func (i *goIpsetProvider) NewIpset(name string, ipsetType string, p *ipset.Params) (Ipset, error) {
-	return nil, fmt.Errorf("Not Supported on non linux platforms")
+func (i *ipsetProvider) NewIpset(name string, ipsetType string, p *ipset.Params) (Ipset, error) {
+	return &winIpSet{}, nil
 }
 
 // GetIpset gets the ipset object from the name.
-func (i *goIpsetProvider) GetIpset(name string) Ipset {
-	return nil
+func (i *ipsetProvider) GetIpset(name string) Ipset {
+	return &winIpSet{}
 }
 
 // DestroyAll destroys all the ipsets - it will fail if there are existing references
-func (i *goIpsetProvider) DestroyAll(prefix string) error {
+func (i *ipsetProvider) DestroyAll(prefix string) error {
 	return nil
 }
 
-func (i *goIpsetProvider) ListIPSets() ([]string, error) {
+func (i *ipsetProvider) ListIPSets() ([]string, error) {
 	return []string{}, nil
 }
 
 // NewGoIPsetProvider Return a Go IPSet Provider
 func NewGoIPsetProvider() IpsetProvider {
-	return &goIpsetProvider{}
+	return &ipsetProvider{}
+}
+
+func (w *winIpSet) Add(entry string, timeout int) error {
+	return nil
+}
+
+func (w *winIpSet) AddOption(entry string, option string, timeout int) error {
+	return nil
+}
+
+func (w *winIpSet) Del(entry string) error {
+	return nil
+}
+
+func (w *winIpSet) Destroy() error {
+	return nil
+}
+
+func (w *winIpSet) Flush() error {
+	return nil
+}
+
+func (w *winIpSet) Test(entry string) (bool, error) {
+	return false, nil
 }
