+++ monitor/internal/pod/delete_controller.go
@@ -1,3 +1,5 @@
+// +build linux
+
 package podmonitor
 
 import (
