+++ monitor/internal/pod/monitor.go
@@ -1,3 +1,5 @@
+// +build linux !windows
+
 package podmonitor
 
 import (
