+++ monitor/internal/pod/resync.go
@@ -1,4 +1,4 @@
-// +build linux
+// +build linux !windows
 
 package podmonitor
